+++ src/main.cpp
@@ -13,7 +13,7 @@ struct Function
 {
 	size_t evals = 0;
 
-	double operator()(const Vector& x)
+	Float operator()(const Vector& x)
 	{
 		evals++;
 		const auto x_shift = (x.array() - 1.).matrix();
@@ -25,8 +25,8 @@ struct Function
 template <typename Callable>
 void call(Callable& o)
 {
-	static_assert(std::is_invocable_r_v<double, Callable, Vector>, "Incorrect objective function type");
-	const double result = o(Vector::Ones(10));
+	static_assert(std::is_invocable_r_v<Float, Callable, Vector>, "Incorrect objective function type");
+	const Float result = o(Vector::Ones(10));
 	std::cout << result;
 }
 
@@ -39,7 +39,7 @@ struct Timer
 	{
 		const auto t2 = high_resolution_clock::now();
 		const auto ms_int = duration_cast<milliseconds>(t2 - t1);
-		std::cout << "Time elapsed: " << static_cast<double>(ms_int.count()) / 1000.0 << "s\n";
+		std::cout << "Time elapsed: " << static_cast<Float>(ms_int.count()) / 1000.0 << "s\n";
 	}
 };
